requirements.txt
streamlit
google-generativeai
Pillow
pandas
st-gsheets-connection